deal with services that are request scoped, fixes #47
not sure how many kittens will die for this fix, but the only other alternative would require a ton of code and complexity, aka a compiler pass that somehow would need to determine which services to read out the class from the definition, then store it somewhere. it would also not work with factory services. so ugly hack it is .. :(