Add fos_rest.serializer.serialize_null to control null value serialization
The JMS serializer by default will not serialize keys with a null value.
This new parameter allows control of this behavior via the FOSRestBundle
configuration in app/config/config.yml.
Setting fos_rest.serializer.serialize_null to true will cause the
setSerializeNull() method on the Context object to be called with the
value that is set in the configuration.
Add test case to check that the Context Object has its member
$serializeNull set to the correct value.
Add new parameter to the Resource/doc/configuration-reference.md
documentation file.