reword expression language sentence(s)
In the current documentation it shows the word json in the rendered output due to the fenced code blocks not starting a line.

I'm not sure how this sentence should really be worded or if you really want it look like a snippet you'd place in your composer.json.
If so, then you probably wanna change this in a different way.

I also rewrote the expression language sentence itself, because i had trouble parsing it.