ignore leading underscores in base normalizer
In #1267, a new normalizer class was introduced that ignores underscores
when converting snake case to camel case. However, the same change was
made to the base class. This means that both classes act in exactly the
same way. Furthermore, this change was a BC break as the behavior of the
base normalizer class was changed without fixing an actual bug.