check for definitions before they might be removed
If the `JMSHandlersPass` is registered as a `TYPE_REMOVE` pass, the
definitions it depends on may already be removed before by the built-in
`RemoveUnusedDefinitionsPass`. That compiler pass removes private
services that are not referenced. This becomes critical, for example,
when running a Symfony 4 based application where services will be
private by default.