Fix dependency to sensio/framework-extra-bundle (#1838)
* Fix dependency to sensio/framework-extra-bundle

\FOS\RestBundle\Request\ParameterBag has a dependency to
\Doctrine\Common\Util\ClassUtils which originates from doctrine/common
which in turn a dependency of sensio/framework-extra-bundle but this is
declared as a dev-dependency. This causes an fatal error with the
message 'Class 'Doctrine\Common\Util\ClassUtils' not found'.
Fixes#259

* Revert "Fix dependency to sensio/framework-extra-bundle"

This reverts commit 436fed7e54da402ed89c2ebad71c62d124a8eb95.

* Add doctrine/common as a dependency

\FOS\RestBundle\Request\ParameterBag has a dependency to
\Doctrine\Common\Util\ClassUtils which originates from doctrine/common
which in turn a dependency of sensio/framework-extra-bundle but this is
declared as a dev-dependency. This causes an fatal error with the
message 'Class 'Doctrine\Common\Util\ClassUtils' not found'.
Fixes#259

* Only use Doctrine's ClassUtils when available

Doctrine is a dependency of a devDependency but used unconditionally.
The use of ClassUtils::getClass is only nessesary when Doctrine is
installed so we check if the class exists before using is and fall back
on get_class if this is not the case.

* Fix comment

* Fix CS

* Useless parameter